refactor(core): lift SEP-2243 wire-format helpers from client/ to core/ (closes 464)
Promote the SEP-2243 helpers from package-private in client/ to public API
in a new core/sep2243.go. Server-side validation and decoding (already
landing piecemeal in server/header_validation.go) needs the same pure
spec rules; lifting prevents duplication and keeps one source of truth.

Promoted (all previously private in client/):

  - ExtractMcpHeaderParams (was extractMcpParamHeaders)
  - ValidateMcpHeaderSchema (was validateMcpParamHeaders)
  - McpParamHeaderName (was mcpParamHeaderName)
  - EncodeMcpHeaderValue (was encodeMcpParamHeaderValue)
  - NeedsBase64Encoding (was needsBase64Encoding)
  - DeriveMcpName (was deriveMcpName)
  - McpMethodHeader, McpNameHeader constants

server/header_validation.go drops its local copies of the two header
constants and consumes core.McpMethodHeader / core.McpNameHeader.
server/streamable_stateless.go updates a single consumer too.

client/headers.go keeps thin one-line wrappers so existing call sites
in client/client.go stay byte-identical; setSEP2243RoutingHeaders
remains client-side (it mutates an outbound *http.Request, not pure
spec).

Pure refactor — `make testconf-upstream-audit` regenerates a
byte-identical UPSTREAM_AUDIT.md.
